requirements.txt
langchain
openai
langchain-openai
streamlit
python-dotenv
streamlit-webrtc
opencv-python-headless
PyGithub